+++ packages/expo-router/types/jest.d.ts
@@ -3,6 +3,7 @@ declare global {
   namespace jest {
     interface Matchers<R> {
       toHavePathname(pathname: string): R;
+      toHavePathnameWithParams(pathname: string): R;
       toHaveSegments(segments: string[]): R;
       toHaveSearchParams(params: Record<string, string>): R;
     }